add type definitions path to package.json
This change involves adding a path to the type declaration file ('.d.ts') for the 'urlcat' module in the package.json file. Typically, TypeScript utilizes this path to ensure type safety when working with the module. This edit is necessary because TypeScript is unable to locate and utilize the type declaration file, even if it exists. This could be due to a mismatch in the paths specified in the "exports" section of the package.json file of the 'urlcat' module. As a result of this change, TypeScript will be able to correctly utilize the types from 'urlcat' in the project. It's possible that the 'urlcat' library may also need to update its package.json or type declaration files to address this issue in the future.